fixes #52 elevation gain data type

+++ src/main/java/org/jstrava/entities/Activity.java
@@ -638,7 +638,7 @@ public class Activity{
 		this.totalElevationGain = totalElevationGain;
 	}
 
-	public int getTotalElevationGain(){
+	public double getTotalElevationGain(){
 		return totalElevationGain;
 	}
